+++ EconomyEX/Systems/D1_RentAdjustSystemMod.cs
@@ -1,4 +1,4 @@
-﻿using Colossal.Entities;
+using Colossal.Entities;
 using Game.Agents;
 using Game.Areas;
 using Game.Buildings;
@@ -19,11 +19,10 @@ using Unity.Collections;
 using Unity.Entities;
 using Unity.Jobs;
 using Unity.Mathematics;
+using EconomyEX.Helpers;
 using Game;
 using Game.Simulation;
 using Transform = Game.Objects.Transform;
-
-using EconomyEX.Helpers;
 
 namespace EconomyEX.Systems
 {